added auth and file upload

+++ app/Http/Controllers/ReviewController.php
@@ -6,11 +6,12 @@ use App\Models\Review;
 use Illuminate\Http\Request;
 use App\Services\BookService;
 use App\Services\ReviewService;
+use Illuminate\Support\Facades\Gate;
 
 
 class ReviewController extends Controller
 {
-    public function __construct(private ReviewService $Service, private BookService $bookService) {}
+    public function __construct(private BookService $service) {}
 
     /**
      * Display a listing of the resource.
@@ -23,17 +24,23 @@ class ReviewController extends Controller
      * Show the form for creating a new review for specified book.
      * DI will auto-create a new Review model for population
      */
-    public function create($id, Review $review)
-    {
+    public function create($id, Review $review, Request $request) { 
+
+        // if (auth()->user()->cannot('create', Review::class)) {
+        // if ($request->user()->cannot('create', Review::class)) {
+        //    return redirect()->back()->with('warning', 'Not authorised'); 
+        //}
         
-        $book = $this->bookService->find($id);
+        $this->authorize('create', Review::class);
+       
+        $book = $this->service->find($id);
         if (!isset($book)) {
-            return redirect()->route('book.index')->with('warning', "Book {$id} does not exist!");
+            return redirect()->route('books.index')->with('warning', "Book {$id} does not exist!");
         }
         
         $review->book_id = $id;
         $review->user_id = auth()->user()->id;
-        return view('reviews.create', ['review' => $review, 'book' => $book]);
+        return view('review.create', ['review' => $review, 'book' => $book]);
     }
 
     /**
@@ -41,17 +48,18 @@ class ReviewController extends Controller
      */
     public function store(Request $request)
     {
+        //$this->authorize('create');
+
         //$request->merge(['reviewed_on' => now()]);
         $review = $request->validate([
             'book_id' => ['required'],
-            'user_id' => ['required'],
-            //'name' => ['required'],            
+            'user_id' => ['required'],           
             'rating' => ['required', 'numeric', 'min:0', 'max:5'],
-            'comment' => ['required','min:5', 'max:1000'],
-            //'reviewed_on' => ['required']
+            'comment' => ['required','min:5', 'max:1000'],            
         ]);
-        $review['reviewed_on'] = now();
-        $review = Review::create($review);       
+    
+        $review = $this->service->addReview($review['book_id'], $review);  
+             
         return redirect()->route("books.show", ['id'=>$review->book_id])
                          ->with('success', "Review Created Successfully");
 
@@ -60,9 +68,14 @@ class ReviewController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Review $review)
+    public function show(int $id)
     {       
-        return view ('reviews.show', ['review' => $review]);
+        $review = Review::find($id);
+        if (!$review) {
+            return redirect()->route("books.index")
+                             ->with('warning', "Review does not exist");
+        }
+        return view ('review.show', ['review' => $review]);
     }
 
     /**
@@ -84,12 +97,17 @@ class ReviewController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Review $review)
+    public function destroy(int $id)
     {
-       $book_id = $review->book_id;
+        //$book = $this->service->deleteReview($id);
 
-       $review->delete();
-       return redirect()->route("books.show", ['id'=>$book_id])
-                        ->with('success', "Review Destroyed Successfully");
-    }
+        $review = Review::find($id);
+        if (!$review) {
+            return redirect()->route("books.index" )->with('error', "Review Not Found");
+        }
+        $book = $review->book;
+        $review->delete();
+        return redirect()->route("books.show", ['id'=>$book->id])
+                         ->with('success', "Review Destroyed Successfully");   
+       }
 }
